mhook_inject: type TimeoutMs as DWORD
DWORD is the conventional type for a millisecond timeout across the Win32
wait APIs (WaitForSingleObject, INFINITE, etc.); use it for
MHOOK_INJECT_PARAMS.TimeoutMs and the internal watcher structs instead of
ULONG. Type-identical (both unsigned long), so no behavioural or layout
change. The MHOOK_PROXY_BLOCK wire field stays ULONG to keep that
serialized struct uniformly fixed-width.

+++ src/mhook-inject/mhook_inject.cpp
@@ -595,7 +595,7 @@ typedef struct _MHOOK_INJECT_WATCHER {
     PVOID   ApcRoutine;        // caller VA (NULL if unused)
     PVOID   ApcContext;
     HANDLE  UserEvent;         // caller's Event (NULL if unused)
-    ULONG   TimeoutMs;
+    DWORD   TimeoutMs;
 } MHOOK_INJECT_WATCHER;
 
 static VOID NTAPI InjectWatcherProc(PVOID arg)
@@ -749,7 +749,7 @@ typedef struct _MHOOK_PROXY_WATCHER {
     PVOID   ApcRoutine;
     PVOID   ApcContext;
     HANDLE  UserEvent;
-    ULONG   TimeoutMs;
+    DWORD   TimeoutMs;
 } MHOOK_PROXY_WATCHER;
 
 static VOID NTAPI ProxyWatcherProc(PVOID arg)
